blog/requirements.txt
asgiref==3.7.2
Django==5.0.2
django-etc==1.4.0
django-hitcount==1.3.5
Pillow==10.1.0
sqlparse==0.4.4
tzdata==2024.1
